Break Digitizer structs and enums into separate header (DigitizerDefs.h) and begin to implement synchronization. Synchronization incomplete (will not compile)
